arxiv_chat/requirements.txt
langchain
langchain-community
faiss-cpu
arxiv
pymupdf
langchain-openai
python-dotenv
streamlit 